Start cleaning things up by splitting the code generator into multiple files.
Currently, we build under cintpos and there is no linker so we simply
include the source into the main code generator file. Eventually, when
we have a stand-alone compiler that we can link, these will become
separate object decks. At that point some scoping issues will need
to be dealt with.