Fix cg_switchon to create an empty basic block after the switch
In LLVM, the switch node is a basic block terminator because it has
branches for all cases, including a default. We can't emit any more
code into the current basic block. However, SYN and TRN can often
emit "catch all" code (typically looking like LN 0 FNRN) which will
break this rule. To deal with this we create a new basic block
switchon.dead, into which such code can be emitted. As this basic
block can't be reached, we'll eliminate it later.