Fix left and right shift for cases undefined in LLVM
The proposed standard (Middleton et al) says for E1 shift E2 "If the value of E2 is an
integer larger than the word length of the implementation then the result of the shift
operation is undefined.". It doesn't say what happens in the defined case of equality.

In Martin Richards' Cintpos manual, it says the result is always defined and is zero for
such shifts.

Alas, this means we can't simply use the instrinsic LLVM shift operations alone, we need
to generate code to spot the case E2 >= BITSPERWORD.