Fix bug provoked by sysb/blib/write_format - unused label
When we generate an IBR, we included labels for parts of code that were
not referenced (that is, LAB Ln create Ln but no other OCODE operation
referenced it). This meant we had LLVM CFG errors.

Mark labels USED when something other then a LAB operations gets their
basic blocks or statics