Reworking of global vector mechanism to replace elaboration with LLVM alias mechanism
When progressing OCODE declaring a global, retrieve the function
associated with the label and create an alias __bcpl_gv<global>
to the name of the function.

In the definition of the global vector, define a weak symbol for each
entry of the form __bcpl_gv<global>.

When the code generator creates this weak symbol as an alias for the
function, the linker will resolve it, placing the function's entry
point in the global.

This assumes that the only items named using the GLOBAL ocode operation
are functions. Other global values are variables and these are only
accessed by their global number vua the LLG, LG and SG ocode operations.