Modifier regression test cmpltest
According to "The BCPL Cintsys and Cintpos User Guide" by
Martin Richards, in a local LET declarion

"The variables N1,...,Nn are allocated consecutive locations in the stack
frame of the current function and so, for instance, the variable Ni may
be accessed by the expression (@N1)!(i − 1).

This feature is a recent addition to the language."

The compiler doesn't honour this requirement at the moment. The feature
isn't mentioned in the "Proposed Definition...".

Modify the test cases for SLCT so that they don't rely on this feature.