Fix a bug in the generation of stack temporaries provided by GOTO
GOTO results in phi node(s) being generated where we decorate them
with all the possible destinations in scope. This causes problems
if we have some alloca temporaries not in the first basic block as
it means the verifier can find control paths through the graph to
uses not dominated by the allocation block.

The fix is to always add the alloca to the end of the first basic
block in the function.