Fix a bug in the use of the A register in RES/RSTACK
We can't declare A (a local alloca) immediately because we may be also
declaring a nested function with its own A. Declare it when the first
RES or RSTACK operation needs it.

Additionally, make sure we declare it in the function's first basic
block. Refactor this process as allocate_temporary as it's common
with the allocation of simulator stack entries.