Fix TRN's assign to remove assumptions about ordering of locals
We don't yet force consecutive locals to be allocated consecutive
addresses in memory. assign() makes this assumption so that it can
build a temporary node in locals.

Add a function splitload that calls load but first places all of
the node components in a VEC we can pass to load.