Rework the simulated stack to handle large VECs better
Instrucduce a mapping from address to value for known elements on
the simulated stack rather than all elements. This way, the size
of a VEC or TABLE has no effect on the size of the current stack
frame.

Also, reintroduce tracing of the OCODE opcode which seems to have been
lost. (We get the name from BCPLSYN.B's opname table.)