Update BCPLSYN template to work around LLVM optimisation issue
When we assign a label to a variable and see the label for the first
time, we create a basic block as the static initialised for the static
that holds the label address. This is what happens in places like
formtree() where we set up the label as the targets for the longjump
on detecting a syntax error.

Later, when the label is defined, we attach the basic block to the
function's sequence of basic blocks and start building content into
it.

During the optimisation state, LLVM detects that the label is not a
branch destination in the current function. We end up with the static
initialised being 1 rather than the basic block address. LLVM doesn't
seem to spot that the label is reachable via the variable we stored
in it.

Fix this by adding a GOTO to the end of the formtree function just
after it calls synerr (as that function doesn't return).