Provide a more conservative implementation of level and longjump
We preserve the non-volatile registers on level and restore them
on longjump.

Because the caller provides no space to store this information, we
define a reserved area for storing level() states. Currently, we
provide space for 128 simultaneous longjumps (a fair few are needed
by the recursive descent parser in BCPLSYN for example).

If this implementation seems robust, we'll reimplement it will
dynamic allocation of memory.